修改Image.Group为Image.PriviewGroup，并简化参数
